Add solana-program-runtime crate
