+++ core/src/cost_update_service.rs
@@ -191,7 +191,7 @@ impl CostUpdateService {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use solana_runtime::message_processor::ProgramTiming;
+    use solana_program_runtime::ProgramTiming;
     use solana_sdk::pubkey::Pubkey;
 
     #[test]
