+++ program-runtime/Cargo.toml
@@ -1,0 +1,33 @@
+[package]
+name = "solana-program-runtime"
+version = "1.8.0"
+description = "Solana program runtime"
+authors = ["Solana Maintainers <maintainers@solana.foundation>"]
+repository = "https://github.com/solana-labs/solana"
+license = "Apache-2.0"
+homepage = "https://solana.com/"
+documentation = "https://docs.rs/solana-program-runtime"
+edition = "2018"
+
+[dependencies]
+libc = "0.2.101"
+libloading = "0.7.0"
+log = "0.4.14"
+num-derive = { version = "0.3" }
+num-traits = { version = "0.2" }
+serde = { version = "1.0.129", features = ["derive", "rc"] }
+solana-frozen-abi = { path = "../frozen-abi", version = "=1.8.0" }
+solana-frozen-abi-macro = { path = "../frozen-abi/macro", version = "=1.8.0" }
+solana-logger = { path = "../logger", version = "=1.8.0" }
+solana-sdk = { path = "../sdk", version = "=1.8.0" }
+thiserror = "1.0"
+
+[lib]
+crate-type = ["lib"]
+name = "solana_program_runtime"
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
+
+[build-dependencies]
+rustc_version = "0.4"
