+++ program-runtime/build.rs
@@ -1,0 +1,1 @@
+../frozen-abi/build.rs
