+++ program-runtime/src/native_loader.rs
@@ -5,6 +5,7 @@ use libloading::os::unix::*;
 use libloading::os::windows::*;
 use log::*;
 use num_derive::{FromPrimitive, ToPrimitive};
+use serde::Serialize;
 use solana_sdk::{
     account::ReadableAccount,
     decode_error::DecodeError,
