+++ program-test/Cargo.toml
@@ -21,6 +21,7 @@ solana-banks-client = { path = "../banks-client", version = "=1.8.0" }
 solana-banks-server = { path = "../banks-server", version = "=1.8.0" }
 solana-bpf-loader-program = { path = "../programs/bpf_loader", version = "=1.8.0" }
 solana-logger = { path = "../logger", version = "=1.8.0" }
+solana-program-runtime = { path = "../program-runtime", version = "=1.8.0" }
 solana-runtime = { path = "../runtime", version = "=1.8.0" }
 solana-sdk = { path = "../sdk", version = "=1.8.0" }
 solana-vote-program = { path = "../programs/vote", version = "=1.8.0" }
