+++ programs/bpf_loader/Cargo.toml
@@ -22,7 +22,7 @@ rand_core = "0.6.3"
 libsecp256k1 = "0.6.0"
 sha3 = "0.9.1"
 solana-measure = { path = "../../measure", version = "=1.8.0" }
-solana-runtime = { path = "../../runtime", version = "=1.8.0" }
+solana-program-runtime = { path = "../../program-runtime", version = "=1.8.0" }
 solana-sdk = { path = "../../sdk", version = "=1.8.0" }
 solana_rbpf = "=0.2.14"
 thiserror = "1.0"
@@ -30,6 +30,7 @@ thiserror = "1.0"
 [dev-dependencies]
 rand = "0.7.3"
 rustversion = "1.0.5"
+solana-runtime = { path = "../../runtime", version = "=1.8.0" }
 
 [lib]
 crate-type = ["lib"]
