+++ runtime/Cargo.toml
@@ -23,8 +23,6 @@ flate2 = "1.0.20"
 fnv = "1.0.7"
 itertools = "0.10.1"
 lazy_static = "1.4.0"
-libc = "0.2.101"
-libloading = "0.7.0"
 log = "0.4.14"
 memmap2 = "0.3.1"
 num-derive = { version = "0.3" }
@@ -43,6 +41,7 @@ solana-frozen-abi-macro = { path = "../frozen-abi/macro", version = "=1.8.0" }
 solana-logger = { path = "../logger", version = "=1.8.0" }
 solana-measure = { path = "../measure", version = "=1.8.0" }
 solana-metrics = { path = "../metrics", version = "=1.8.0" }
+solana-program-runtime = { path = "../program-runtime", version = "=1.8.0" }
 solana-rayon-threadlimit = { path = "../rayon-threadlimit", version = "=1.8.0" }
 solana-sdk = { path = "../sdk", version = "=1.8.0" }
 solana-secp256k1-program = { path = "../programs/secp256k1", version = "=1.8.0" }
