+++ runtime/src/lib.rs
@@ -25,7 +25,6 @@ pub mod instruction_recorder;
 pub mod loader_utils;
 pub mod log_collector;
 pub mod message_processor;
-mod native_loader;
 pub mod neon_evm_program;
 pub mod non_circulating_supply;
 mod pubkey_bins;
